lab 18

+++ 18_jpa_hibernate/task_2/src/main/java/com/nixsolutions/dao/RoleDao.java
@@ -6,7 +6,7 @@ import com.nixsolutions.entity.Role;
 
 public interface RoleDao {
 	public List<Role> getAllRoles();
-	public Role getRoleById(int roleId);
+	public Role getRoleById(Long roleId);
 	public void createRole(Role role);
 	public void updateRole(Role role);
 	public void deleteRole(Role role);
